# open-api-spec/src/main/resources/META-INF/resources/openapi/fx-trading-systems-api.yaml
openapi: 3.0.3
info:
  title: FX Trading Systems API
  version: 2.5.0
  description: |
    Contract for requesting executable FX prices and booking trades through a trading system.
    A one-way request returns one flat executable price; a two-way request returns flat BUY and SELL fields.
    Quotes are single-use and expire at `expiresAt`. All rates are quote currency per base currency.
    Identification is business context for this local trading system, not authentication.
    POST bodies carry requestId/channel/segment/customerId. GET operations carry the equivalent X-* headers.
    Responses echo the current request context; originalRequestId identifies the resource's creation request.
    State and idempotency are bounded, in-memory, single-instance, and reset on restart.
    Default quote retention is five minutes after expiry; trade and idempotency retention are 24 hours.
    Capacity exhaustion returns 503 without evicting retained records or consuming a quote.
    Resting orders rest until triggered, expired, or cancelled, and are evaluated on a fixed interval.
    A limit order is the only resting order type today; its limitPrice is what ends the wait.
    A terminal order is reported to the caller's callbackUrl at least once; receivers dedupe on eventId.
servers:
  - url: http://localhost:8090
    description: Local trading system
...
paths:
...
  /api/v1/resting-orders:
    post:
      tags: [Resting Orders]
      operationId: placeRestingOrder
      summary: Place a resting order
      description: |
        The order rests until its limit is reachable, its good-till time passes, or it is cancelled.
        The caller names the order with orderId, and that name is also the idempotency key: this endpoint
        takes no Idempotency-Key header. Repeating a placement with the same orderId and identical terms
        returns 200 and the original order, so a lost response is safe to retry. Reusing the orderId for
        different terms returns 409, for as long as that order is retained; afterwards the name is free.
        The name is scoped by customerId, so two customers may use the same one.
        GOOD_TILL_TIME requires expiresAt in the future; GOOD_TILL_CANCELLED must omit it.
        Working orders are evaluated on a fixed interval, so triggering is not instantaneous and the
        executed price is the one quoted at the evaluation that triggered it, never the limit itself.
        A triggered order books a trade retrievable at /api/v1/bookings/{tradeId}, then the trading system
        delivers a callback to callbackUrl. Orders and their events are bounded in-memory state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestingOrderRequest'
      responses:
        '201':
          description: Resting order accepted and working.
          headers:
            Location:
              description: URI of the created order.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestingOrder'
        '200':
          description: The original order returned for an idempotent replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestingOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/CapacityExceeded'
        default:
          $ref: '#/components/responses/Problem'
      callbacks:
        restingOrderEvent:
          '{$request.body#/callbackUrl}':
            post:
              tags: [Resting Orders]
              operationId: onRestingOrderEvent
              summary: Terminal-state notification delivered by the trading system
              description: |
                Delivered at least once when a working order becomes TRIGGERED or EXPIRED.
                eventId is stable across retries of the same event, so receivers must dedupe on it.
                Events for one order are delivered in order; delivery never blocks order evaluation.
                Any 2xx completes delivery. Other responses and transport failures are retried with
                exponential backoff until the attempt limit, after which callbackStatus becomes FAILED.
                Cancellations are not delivered: the caller already knows the outcome from the response.
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/RestingOrderEvent'
              responses:
                '200':
                  description: Event accepted.
                '204':
                  description: Event accepted.
                default:
                  description: Delivery is retried until the attempt limit is reached.
  /api/v1/resting-orders/{orderId}:
    get:
      tags: [Resting Orders]
      operationId: getRestingOrder
      summary: Get the current state of a resting order
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/ChannelHeader'
        - $ref: '#/components/parameters/SegmentHeader'
        - $ref: '#/components/parameters/CustomerIdHeader'
      responses:
        '200':
          description: Order found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestingOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Resting Orders]
      operationId: cancelRestingOrder
      summary: Cancel a working resting order
      description: |
        Only a WORKING order can be cancelled; an order that already reached a terminal state returns 409.
        Cancellation races evaluation: an order that triggers first stays TRIGGERED and its callback is still delivered.
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/ChannelHeader'
        - $ref: '#/components/parameters/SegmentHeader'
        - $ref: '#/components/parameters/CustomerIdHeader'
      responses:
        '200':
          description: Order cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestingOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Problem'
    put:
      tags: [Resting Orders]
      operationId: amendRestingOrder
      summary: Amend a working resting order
      description: |
        Atomically replaces the mutable terms of a WORKING order while preserving its orderId, original
        request identity, instrument, side, callback URL, and placement timestamp. Quantity, limit price,
        and time in force are replaced together. The latest evaluation is cleared because it described the
        previous terms. A terminal order returns 409.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestingOrderAmendRequest'
      responses:
        '200':
          description: Working order amended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestingOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Problem'
components:
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      description: The caller-assigned order identifier.
      schema:
        $ref: '#/components/schemas/OrderId'
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/RequestId'
    ChannelHeader:
      name: X-Channel
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Channel'
    SegmentHeader:
      name: X-Segment
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Segment'
    CustomerIdHeader:
      name: X-Customer-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/CustomerId'
...
  responses:
    BadRequest:
      description: The request is malformed or violates the contract.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    UnprocessableEntity:
      description: The request is valid but the instrument is unsupported or the callback URL is not allowed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    NotFound:
      description: The requested quote, trade, or order does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    Conflict:
      description: |
        The quote was already booked, the order is no longer working, the orderId already names a
        different order, or a booking idempotency key was reused inconsistently.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
...
    CapacityExceeded:
      description: Retained state is at capacity. Retry later with the same booking idempotency key.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    Problem:
      description: Request could not be completed; available request context is echoed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
  schemas:
...
    RestingOrderRequest:
      type: object
      additionalProperties: false
      description: |
        side and limitPrice describe quantityCurrency from the customer's perspective, in the currencyPair's
        orientation. A BUY triggers when the executable client price is at or below limitPrice; a SELL triggers
        when it is at or above. A limit that is already reachable triggers at the first evaluation after placement.
        GOOD_TILL_TIME requires expiresAt strictly in the future; GOOD_TILL_CANCELLED must omit expiresAt.
      required: [requestId, orderId, channel, segment, customerId, currencyPair, quantity, quantityCurrency, tenor, side, limitPrice, timeInForce, callbackUrl]
      properties:
        requestId:
          $ref: '#/components/schemas/RequestId'
        orderId:
          $ref: '#/components/schemas/OrderId'
        channel:
          $ref: '#/components/schemas/Channel'
        segment:
          $ref: '#/components/schemas/Segment'
        customerId:
          $ref: '#/components/schemas/CustomerId'
        currencyPair:
          $ref: '#/components/schemas/CurrencyPair'
        quantity:
          $ref: '#/components/schemas/Quantity'
        quantityCurrency:
          $ref: '#/components/schemas/Currency'
        tenor:
          $ref: '#/components/schemas/Tenor'
        side:
          $ref: '#/components/schemas/Side'
        limitPrice:
          $ref: '#/components/schemas/Rate'
        timeInForce:
          $ref: '#/components/schemas/TimeInForce'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        callbackUrl:
          $ref: '#/components/schemas/CallbackUrl'
    RestingOrderAmendRequest:
      type: object
      additionalProperties: false
      description: |
        Complete replacement for a working order's mutable terms. Identification must match the order.
        GOOD_TILL_TIME requires expiresAt strictly in the future; GOOD_TILL_CANCELLED must omit it.
      required: [requestId, channel, segment, customerId, quantity, limitPrice, timeInForce]
      properties:
        requestId:
          $ref: '#/components/schemas/RequestId'
        channel:
          $ref: '#/components/schemas/Channel'
        segment:
          $ref: '#/components/schemas/Segment'
        customerId:
          $ref: '#/components/schemas/CustomerId'
        quantity:
          $ref: '#/components/schemas/Quantity'
        limitPrice:
          $ref: '#/components/schemas/Rate'
        timeInForce:
          $ref: '#/components/schemas/TimeInForce'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
    RestingOrder:
      type: object
      additionalProperties: false
      description: |
        lastEvaluatedAt and lastEvaluatedPrice report the most recent evaluation, so a working order shows
        why it has not triggered, and closedAt is set once the order leaves WORKING. The order carries no
        execution price detail: a TRIGGERED order's fill is described by its callback event and by the trade
        identified by tradeId. Exposing that identifier lets a receiver reconcile safely after downtime.
      required: [requestId, channel, segment, customerId, originalRequestId, responseId, responseAt, orderId, currencyPair, quantity, quantityCurrency, tenor, side, limitPrice, timeInForce, status, placedAt, callbackUrl, callbackStatus, callbackAttempts]
      properties:
        requestId:
          $ref: '#/components/schemas/RequestId'
        channel:
          $ref: '#/components/schemas/Channel'
        segment:
          $ref: '#/components/schemas/Segment'
        customerId:
          $ref: '#/components/schemas/CustomerId'
        originalRequestId:
          $ref: '#/components/schemas/RequestId'
        responseId:
          $ref: '#/components/schemas/Identifier'
        responseAt:
          $ref: '#/components/schemas/Timestamp'
        orderId:
          $ref: '#/components/schemas/OrderId'
        currencyPair:
          $ref: '#/components/schemas/CurrencyPair'
        quantity:
          $ref: '#/components/schemas/Quantity'
        quantityCurrency:
          $ref: '#/components/schemas/Currency'
        tenor:
          $ref: '#/components/schemas/Tenor'
        side:
          $ref: '#/components/schemas/Side'
        limitPrice:
          $ref: '#/components/schemas/Rate'
        timeInForce:
          $ref: '#/components/schemas/TimeInForce'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        status:
          $ref: '#/components/schemas/RestingOrderStatus'
        placedAt:
          $ref: '#/components/schemas/Timestamp'
        lastEvaluatedAt:
          $ref: '#/components/schemas/Timestamp'
        lastEvaluatedPrice:
          $ref: '#/components/schemas/Rate'
        closedAt:
          $ref: '#/components/schemas/Timestamp'
        tradeId:
          $ref: '#/components/schemas/Identifier'
        callbackUrl:
          $ref: '#/components/schemas/CallbackUrl'
        callbackStatus:
          $ref: '#/components/schemas/CallbackStatus'
        callbackAttempts:
          $ref: '#/components/schemas/AttemptCount'
    RestingOrderEvent:
      oneOf:
        - $ref: '#/components/schemas/RestingOrderTriggeredEvent'
        - $ref: '#/components/schemas/RestingOrderExpiredEvent'
      discriminator:
        propertyName: eventType
        mapping:
          TRIGGERED: '#/components/schemas/RestingOrderTriggeredEvent'
          EXPIRED: '#/components/schemas/RestingOrderExpiredEvent'
    RestingOrderTriggeredEvent:
      type: object
      additionalProperties: false
      description: |
        The order matched and a trade was booked. The trade is retrievable at /api/v1/bookings/{tradeId}
        using this event's channel, segment, and customerId. Settlement and price fields repeat the booked
        trade exactly, so a receiver never has to call back to act on the fill.
      required: [eventId, eventType, occurredAt, attempt, orderId, originalRequestId, channel, segment, customerId, currencyPair, quantity, quantityCurrency, tenor, side, limitPrice, status, tradeId, coverPrice, clientPrice, swapPoints, buyCurrency, buyQuantity, sellCurrency, sellQuantity, spotDate, valueDate]
      properties:
        eventId:
          $ref: '#/components/schemas/Identifier'
        eventType:
          type: string
          pattern: '^TRIGGERED$'
          description: Must be TRIGGERED; selects this contract variant.
        occurredAt:
          $ref: '#/components/schemas/Timestamp'
        attempt:
          $ref: '#/components/schemas/AttemptCount'
        orderId:
          $ref: '#/components/schemas/OrderId'
        originalRequestId:
          $ref: '#/components/schemas/RequestId'
        channel:
          $ref: '#/components/schemas/Channel'
        segment:
          $ref: '#/components/schemas/Segment'
        customerId:
          $ref: '#/components/schemas/CustomerId'
        currencyPair:
          $ref: '#/components/schemas/CurrencyPair'
        quantity:
          $ref: '#/components/schemas/Quantity'
        quantityCurrency:
          $ref: '#/components/schemas/Currency'
        tenor:
          $ref: '#/components/schemas/Tenor'
        side:
          $ref: '#/components/schemas/Side'
        limitPrice:
          $ref: '#/components/schemas/Rate'
        status:
          $ref: '#/components/schemas/RestingOrderStatus'
        tradeId:
          $ref: '#/components/schemas/Identifier'
        coverPrice:
          $ref: '#/components/schemas/Rate'
        clientPrice:
          $ref: '#/components/schemas/Rate'
        swapPoints:
          $ref: '#/components/schemas/SwapPoints'
        buyCurrency:
          $ref: '#/components/schemas/Currency'
        buyQuantity:
          $ref: '#/components/schemas/SettlementQuantity'
        sellCurrency:
          $ref: '#/components/schemas/Currency'
        sellQuantity:
          $ref: '#/components/schemas/SettlementQuantity'
        spotDate:
          $ref: '#/components/schemas/Date'
        valueDate:
          $ref: '#/components/schemas/Date'
    RestingOrderExpiredEvent:
      type: object
      additionalProperties: false
      description: The good-till time passed while the order was still working. No trade exists.
      required: [eventId, eventType, occurredAt, attempt, orderId, originalRequestId, channel, segment, customerId, currencyPair, quantity, quantityCurrency, tenor, side, limitPrice, status, expiresAt]
      properties:
        eventId:
          $ref: '#/components/schemas/Identifier'
        eventType:
          type: string
          pattern: '^EXPIRED$'
          description: Must be EXPIRED; selects this contract variant.
        occurredAt:
          $ref: '#/components/schemas/Timestamp'
        attempt:
          $ref: '#/components/schemas/AttemptCount'
        orderId:
          $ref: '#/components/schemas/OrderId'
        originalRequestId:
          $ref: '#/components/schemas/RequestId'
        channel:
          $ref: '#/components/schemas/Channel'
        segment:
          $ref: '#/components/schemas/Segment'
        customerId:
          $ref: '#/components/schemas/CustomerId'
        currencyPair:
          $ref: '#/components/schemas/CurrencyPair'
        quantity:
          $ref: '#/components/schemas/Quantity'
        quantityCurrency:
          $ref: '#/components/schemas/Currency'
        tenor:
          $ref: '#/components/schemas/Tenor'
        side:
          $ref: '#/components/schemas/Side'
        limitPrice:
          $ref: '#/components/schemas/Rate'
        status:
          $ref: '#/components/schemas/RestingOrderStatus'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        lastEvaluatedPrice:
          $ref: '#/components/schemas/Rate'
    ApiProblem:
      type: object
      additionalProperties: false
      description: Identification fields are echoed when available, including invalid supplied strings.
      required: [responseId, responseAt, type, title, status, detail, errorCode]
      properties:
        responseId:
          $ref: '#/components/schemas/Identifier'
        responseAt:
          $ref: '#/components/schemas/Timestamp'
        requestId:
          type: string
        channel:
          type: string
        segment:
          type: string
        customerId:
          type: string
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        errorCode:
          type: string
    RequestId:
      type: string
      minLength: 1
      maxLength: 100
      pattern: '^[A-Za-z0-9._:-]+$'
    Channel:
      type: string
      minLength: 1
      maxLength: 20
      pattern: '^\S+$'
      example: WEB
    Segment:
      type: string
      minLength: 1
      maxLength: 1
      pattern: '^\S$'
      example: C
    CustomerId:
      type: string
      minLength: 10
      maxLength: 10
      pattern: '^\S{10}$'
      example: '0000123456'
    CurrencyPair:
      type: string
      pattern: '^[A-Z]{6}$'
      example: EURUSD
      description: Base currency followed by quote currency. All prices use this orientation.
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
    Quantity:
      type: number
      minimum: 0
      exclusiveMinimum: true
      maximum: 1000000000000
      description: |
        Requested amount in quantityCurrency, which must be one leg of currencyPair.
        Must fit currency minor units (JPY zero decimals; other supported currencies two).
        Preserved exactly at booking. Notional size tiers use spot-equivalent base-currency quantity.
    SettlementQuantity:
      type: number
      minimum: 0
      exclusiveMinimum: true
      description: |
        Positive customer settlement amount. The requested leg is preserved exactly.
        The counter leg is calculated using clientPrice (quote currency per base currency):
        multiply a base amount or divide a quote amount, then round once HALF_UP to the
        counter currency's minor units (JPY zero decimals; other supported currencies two).
        Requests whose counter amount rounds to zero are rejected during pricing.
    Rate:
      type: number
      minimum: 0
      exclusiveMinimum: true
      description: |
        All-in outright rate in quote-currency units per one base-currency unit.
        Cover price is before markup; client price is the executable customer rate.
        Rounded HALF_UP to three decimals for JPY quote currency, five otherwise.
    SwapPoints:
      type: number
      description: Forward minus spot mid in outright-rate units, not pips; zero for SPOT.
    CallbackUrl:
      type: string
      maxLength: 2000
      pattern: '^https?://\S+$'
      example: http://localhost:8080/api/resting-orders/events
      description: |
        Absolute http(s) URL the trading system POSTs order events to. It must start with one of the
        trading system's configured allowed prefixes; any other URL is rejected with 422.
    AttemptCount:
      type: integer
      format: int32
      minimum: 0
      description: Delivery attempts made so far, counting the first attempt as one.
    OrderId:
      type: string
      minLength: 1
      maxLength: 100
      pattern: '^[A-Za-z0-9._:-]+$'
      example: ORD-20260908-000123
      description: |
        Caller-assigned name for the order, and the idempotency key for placing it. Unique per customerId
        for as long as the order is retained: repeating identical terms replays the original order, while
        different terms are rejected. Two customers may use the same value.
    Identifier:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    Date:
      type: string
      format: date
    Tenor:
      type: string
      enum: [SPOT, ONE_WEEK, ONE_MONTH, THREE_MONTHS, SIX_MONTHS, ONE_YEAR]
      description: |
        UTC trade date; configurable pair spot lag (default T+2, USDCAD T+1), weekends and
        configured currency holidays excluded. Forward dates preserve month-end and use
        modified following. Synthetic currency curves accrue ACT/365F over spotDate to valueDate.
    Side:
      type: string
      description: BUY or SELL quantityCurrency from the customer's perspective.
      enum: [BUY, SELL]
    TimeInForce:
      type: string
      enum: [GOOD_TILL_CANCELLED, GOOD_TILL_TIME]
      description: GOOD_TILL_TIME requires expiresAt; GOOD_TILL_CANCELLED must omit it.
    RestingOrderStatus:
      type: string
      enum: [WORKING, TRIGGERED, EXPIRED, CANCELLED]
      description: |
        WORKING orders are evaluated on the trading system's interval. TRIGGERED, EXPIRED, and
        CANCELLED are terminal; only TRIGGERED and EXPIRED produce a callback.
    CallbackStatus:
      type: string
      enum: [NOT_REQUIRED, PENDING, DELIVERED, FAILED]
      description: |
        NOT_REQUIRED while an order is working and for cancellations. PENDING once an event is
        queued or awaiting retry, DELIVERED after a 2xx, FAILED once attempts are exhausted.
...